Phi-3-chat/requirements.txt
azure-search-documents==11.6.0b4
azure-storage-blob==12.20.0
azure-identity==1.16.0
azure-keyvault-secrets==4.3.0
azure-ai-ml
aiohttp
ipykernel
python-dotenv
requests
openai